WIP:  split functions into logical modules.  Wiring is not complete. just a cut/paste into new files
